feat: add ViewChange Injector.
