+++ growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/inject/ViewClickInjector.java
@@ -18,11 +18,8 @@ package com.growingio.android.sdk.autotrack.inject;
 
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.CompoundButton;
 import android.widget.ExpandableListView;
 import android.widget.RadioGroup;
-import android.widget.RatingBar;
-import android.widget.SeekBar;
 import android.widget.Spinner;
 
 import com.google.android.material.button.MaterialButtonToggleGroup;
@@ -36,10 +33,6 @@ public class ViewClickInjector {
 
     public static void viewOnClick(View.OnClickListener listener, View view) {
         ViewClickProvider.viewOnClickListener(view);
-    }
-
-    public static void seekBarOnSeekBarChange(SeekBar.OnSeekBarChangeListener listener, SeekBar seekBar) {
-        ViewClickProvider.seekBarOnClick(seekBar);
     }
 
     public static void adapterViewOnItemClick(AdapterView.OnItemClickListener listener, AdapterView adapterView, View view, int position, long id) {
@@ -61,11 +54,6 @@ public class ViewClickInjector {
         ViewClickProvider.expandableListViewOnChildClick(v);
     }
 
-
-    public static void compoundButtonOnChecked(CompoundButton.OnCheckedChangeListener listener, CompoundButton button, boolean checked) {
-        ViewClickProvider.compoundButtonOnCheck(button);
-    }
-
     public static void radioGroupOnChecked(RadioGroup.OnCheckedChangeListener listener, RadioGroup radioGroup, int i) {
         ViewClickProvider.radioGroupOnCheck(radioGroup.findViewById(radioGroup.getCheckedRadioButtonId()));
     }
@@ -73,12 +61,6 @@ public class ViewClickInjector {
     public static void materialButtonToggleGroupOnButtonChecked(MaterialButtonToggleGroup.OnButtonCheckedListener listener, MaterialButtonToggleGroup group, int checkedId, boolean isChecked) {
         if (isChecked) {
             ViewClickProvider.materialButtonToggleGroupOnButtonCheck(group.findViewById(checkedId));
-        }
-    }
-
-    public static void ratingBarOnRatingBarChange(RatingBar.OnRatingBarChangeListener listener, RatingBar ratingBar, float rating, boolean fromUser) {
-        if (fromUser) {
-            ViewClickProvider.ratingBarOnRatingBarChange(ratingBar);
         }
     }
 }
