+++ growingio-autotracker-core/src/test/java/com/growingio/android/sdk/autotrack/UtilTest.java
@@ -25,7 +25,7 @@ import androidx.appcompat.view.menu.ListMenuItemView;
 import androidx.test.core.app.ApplicationProvider;
 
 import com.google.common.truth.Truth;
-import com.growingio.android.sdk.autotrack.change.ViewChangeProvider;
+import com.growingio.android.sdk.autotrack.inject.ViewChangeProvider;
 import com.growingio.android.sdk.autotrack.inject.ActivityInjector;
 import com.growingio.android.sdk.autotrack.shadow.AlertControllerShadow;
 import com.growingio.android.sdk.autotrack.shadow.ListMenuItemViewShadow;
@@ -82,7 +82,7 @@ public class UtilTest {
     public void activityInjectTest() {
         RobolectricActivity activity = Robolectric.buildActivity(RobolectricActivity.class).get();
         ActivityInjector.onActivityNewIntent(activity, new Intent());
-        ViewChangeProvider.viewOnChange(activity.getTextView());
+        ViewChangeProvider.editTextOnFocusChange(activity.getTextView());
     }
 
 }
